ai-server/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
langchain
langgraph
faiss-cpu
lightgbm
xgboost
transformers
torch
pydantic_settings